feat: Add `strict` argument
If `strict=True` (default) only parameters that are explicitly listed in
the schema are allowed, otherwise an error is thrown (useful to make
users aware of typos in the config).
With `strict=False` the old behaviour of silently merging everything, no
matter if it was in the schema or not, is enabled.