Remove TODO
Loading a schema from a file could be a nice feature but would also be
some effort and result in additional code that needs to be maintained.
Follow the YAGNI principle and remove the TODO (can still be implemented
at some point if there is an actual use case for it).